requirements.txt
farm-haystack
streamlit
pandas
argparse
